backend/ingest/requirements.txt
functions-framework==3.*
google-genai>=0.3.0
google-cloud-aiplatform>=1.71.0
